Styled the App, Navigation, and Landing Components
Styled App component so that nothing overflows horizontally and it will grow (height-wise) to the size of its content. Navigation is a flex container holding static text and is position horizontally on top. The Landing Component is made of two sides that is holding coloured divs as placeholders for images and text.