Refactored code
Cleaned up every component file so thatthere are no unnecessary comments. Changed the names on some variables so that they would reflect better what is being stored inside of it. Refactored bloated code to be more dynamic so that it would be easier to maintain and be easier to read.