Styled Product component and finished adding routing
The Product component holds a Link wrapping the innerImage element which takes the user to that product's ProductPage. Added a small CSS hover animation which causes the inner image to grow to full width and height of container. Added routing to the navigation items so that you can now go back Home by the Erased Closet text.