Set up Cart Modal and Backdrop Modal components
Clicking on CART in navigation will open the side drawer Cart Modal. The backdrop will render depending on the state of the Cart. Both accept methods as props to change the state of the Cart modal. Created a basic layout for the landing page that is not yet separated into its own component.