Added async action creators
Installed redux-thunk to be able to create asynchronous action creators. Created async fetchProducts, which sends a request to the server to get all products as a response. The response is passed a syncronous action creator, and dispatched to the Product reducer. Inside App.js, on mount, the action is dispatched, and returns the list of products.