Created new page for creating new products
Create a NewProductPage componentthat is currently rendering a form. The form has custom inputs, created in the shared/components folder. The Input component is handling complex validation logic, and depends on my props to configure itself. Using a reducer and handlers to decide whether the current value is valid, which depends on its configuration.