Created Admin page, fetching products in App.js
The Admin page is where the admin of the site can manage products, users and orders. The App.js file is fetching the ddata from the backen d using asyn action creators, and then stores it in the global redux store. Cleaned up the App.js File and removed unecessary lines of code.