Styled the Admin component
Admin renders the table headers for products and the ProductList component which will create a row in the Admin's main grid layout. Each row has it's own internal grid layout to match the headers. Created a button with the text create. When clicked, a form to add a new product to the database will be shown in a pop up form.