Added add and delete functionality in Admin component
The trash svg icon, will now dispatch an async action to the redux store. It makes a request to the delete route on the backend. The redux store updates the state to reflect changes in the database. The new product form is now hooked up to the redux store as well, sending a post request to the backend during the action object's journey to the reducer. Experimenting with the create product form styles.