Refactored async thunk actions and App component
In actions/product, the fetchProducts function is explicitly set to async so that async/await can be used within it. The saveProducts method was moved into the fetchProducts function. Now the saveProducts function is dispatching the action and returning a response that can be assigned to a variable at component level. In App, useEffect calls the dispatch function and recieves the result of saveProducts.