Connected SignUp component to the backend
Created action creator and reducer to handle users. SignUp will send a POST request to the backend to create a new user based on the information passed. If successful, the reducer updates the current user property to be newly created user, and redirects user to the homepage.The SignUp component is connected to the redux store and is dispatching when the form is submitted.