Added add image functionality to ImageUpload component
When a user selects a file, that file will be change the current state: file, isValid. A fileReader object is created when file changes, and it reads the content of that file and set it to the filePreviewUrl. The image element is rendered to the DOM and the file the user selected will be shown.