Refactored product.js async dispatch functions
Refactored the fetchProducts, fetchProduct, deleteProduct, and createProduct methods. They are now functions that return an async dispatch function allowing the code inside to use async/await and try/catch. Created a utility function isBadHttpResponse which will check if the response sent back to client is 400 - 599, and if true will throw an error, otherwise a function will be called to dispatch an action to the product reducer